% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TTOBRC.R
\docType{data}
\name{dat}
\alias{dat}
\title{Data were extracted from nine Surveillance, Epidemiology, and End Results
registries. It consisted of 7986 women who were diagnosed with HD and
received their primary treatment between 1973 and 1999, and the incidence
of subsequently diagnosed breast cancer.}
\format{\describe{
\item{TTOBRC}{Time to the development of breast cancer, measured in days.}
\item{STATUS}{A binary indicator of the occurrence of breast cancer.}
\item{RT}{A binary indicator of having received radiotherapy treatment.}
\item{AGE}{Subjects' Age}
}}
\source{
\url{http://seer.cancer.gov}
}
\usage{
dat
}
\description{
Downloaded by Guangming Lang.
}
\keyword{datasets}

