% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{mk_box_plt}
\alias{mk_box_plt}
\title{Creates a function that draws boxplots}
\usage{
mk_box_plt(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
a function that takes three arguments:
             xvar     : a string specifying the categorical x variable
             yvar     : a string specifying the continuous y variable
             degree   : a number specifying the angle of x-axis tick labels
             vpos     : a number specifying the additional height of the text
                        labels beyond the max y-value of each group shown on
                        the main graph
             log_y    : a logical value specifying whether to use log10 scale
                        on y-axis. Default is FALSE
}
\description{
Creates a function that draws boxplots
}

