% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/run_descriptive.R
\name{run_descriptive}
\alias{run_descriptive}
\title{Run some descriptive analysis.}
\usage{
run_descriptive()
}
\value{
summary stats and a boxplot
}
\description{
Run some descriptive analysis.
}

